call_analytics/requirements.txt
chromadb==0.6.3
django-extensions==3.2.3
gradio==5.23.3
httptools==0.6.4
jaraco.collections==5.1.0
langchain-community==0.3.20
langchain-groq==0.3.2
matplotlib==3.10.1
peft==0.15.1
pip-chill==1.0.3
platformdirs==4.2.2
reportlab==4.3.1
sentence-transformers==4.0.1
spacy==3.8.5
tomli==2.0.1
watchfiles==1.0.4
